feat(composition): Implement WhiteNoiseEffect + Sample + Win2D + Refactor
