+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/CanvasComposite.cs
@@ -1,0 +1,19 @@
+﻿namespace Microsoft.Graphics.Canvas
+{
+	public enum CanvasComposite
+	{
+		SourceOver = 0,
+		DestinationOver = 1,
+		SourceIn = 2,
+		DestinationIn = 3,
+		SourceOut = 4,
+		DestinationOut = 5,
+		SourceAtop = 6,
+		DestinationAtop = 7,
+		Xor = 8,
+		Add = 9,
+		Copy = 10,
+		BoundedCopy = 11,
+		MaskInvert = 12
+	}
+}
