+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/CanvasEdgeBehavior.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.Graphics.Canvas
+{
+	public enum CanvasEdgeBehavior
+	{
+		Clamp = 0,
+		Wrap = 1,
+		Mirror = 2
+	}
+}
+
