+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/Effects/BlendEffectMode.cs
@@ -1,0 +1,32 @@
+﻿namespace Microsoft.Graphics.Canvas.Effects
+{
+	public enum BlendEffectMode
+	{
+		Multiply = 0,
+		Screen = 1,
+		Darken = 2,
+		Lighten = 3,
+		Dissolve = 4,
+		ColorBurn = 5,
+		LinearBurn = 6,
+		DarkerColor = 7,
+		LighterColor = 8,
+		ColorDodge = 9,
+		LinearDodge = 10,
+		Overlay = 11,
+		SoftLight = 12,
+		HardLight = 13,
+		VividLight = 14,
+		LinearLight = 15,
+		PinLight = 16,
+		HardMix = 17,
+		Difference = 18,
+		Exclusion = 19,
+		Hue = 20,
+		Saturation = 21,
+		Color = 22,
+		Luminosity = 23,
+		Subtract = 24,
+		Division = 25
+	}
+}
