+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/Effects/EffectBorderMode.cs
@@ -1,0 +1,8 @@
+﻿namespace Microsoft.Graphics.Canvas.Effects
+{
+	public enum EffectBorderMode
+	{
+		Soft = 0,
+		Hard = 1
+	}
+}
