+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/Effects/EffectOptimization.cs
@@ -1,0 +1,9 @@
+﻿namespace Microsoft.Graphics.Canvas.Effects
+{
+	public enum EffectOptimization
+	{
+		Speed = 0,
+		Balanced = 1,
+		Quality = 2
+	}
+}
