+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/Effects/PointSpecularEffect.cs
@@ -1,0 +1,92 @@
+﻿using System;
+using System.Numerics;
+using System.Runtime.InteropServices;
+using Windows.Graphics.Effects;
+using Windows.Graphics.Effects.Interop;
+using Windows.UI;
+
+namespace Microsoft.Graphics.Canvas.Effects
+{
+	[Guid("09C3CA26-3AE2-4F09-9EBC-ED3865D53F22")]
+	public class PointSpecularEffect : IGraphicsEffect, IGraphicsEffectSource, IGraphicsEffectD2D1Interop
+	{
+		private string _name = "PointSpecularEffect";
+		private Guid _id = new Guid("09C3CA26-3AE2-4F09-9EBC-ED3865D53F22");
+
+		public string Name
+		{
+			get => _name;
+			set => _name = value;
+		}
+
+		public Vector3 LightPosition { get; set; }
+
+		public float SpecularExponent { get; set; } = 1.0f;
+
+		public float SpecularAmount { get; set; } = 1.0f;
+
+		public Color LightColor { get; set; } = Colors.White;
+
+		public IGraphicsEffectSource Source { get; set; }
+
+		public Guid GetEffectId() => _id;
+
+		public void GetNamedPropertyMapping(string name, out uint index, out GraphicsEffectPropertyMapping mapping)
+		{
+			switch (name)
+			{
+				case "LightPosition":
+					{
+						index = 0;
+						mapping = GraphicsEffectPropertyMapping.Direct;
+						break;
+					}
+				case "SpecularExponent":
+					{
+						index = 1;
+						mapping = GraphicsEffectPropertyMapping.Direct;
+						break;
+					}
+				case "SpecularAmount":
+					{
+						index = 2;
+						mapping = GraphicsEffectPropertyMapping.Direct;
+						break;
+					}
+				case "LightColor":
+					{
+						index = 3;
+						mapping = GraphicsEffectPropertyMapping.ColorToVector3;
+						break;
+					}
+				default:
+					{
+						index = 0xFF;
+						mapping = (GraphicsEffectPropertyMapping)0xFF;
+						break;
+					}
+			}
+		}
+
+		public object GetProperty(uint index)
+		{
+			switch (index)
+			{
+				case 0:
+					return LightPosition;
+				case 1:
+					return SpecularExponent;
+				case 2:
+					return SpecularAmount;
+				case 3:
+					return LightColor;
+				default:
+					return null;
+			}
+		}
+
+		public uint GetPropertyCount() => 4;
+		public IGraphicsEffectSource GetSource(uint index) => Source;
+		public uint GetSourceCount() => 1;
+	}
+}
