+++ src/Uno.UI.Composition/Win2D/Microsoft/Graphics/Canvas/Effects/SepiaEffect.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using Windows.Graphics.Effects;
+using Windows.Graphics.Effects.Interop;
+
+namespace Microsoft.Graphics.Canvas.Effects
+{
+	[Guid("3A1AF410-5F1D-4DBE-84DF-915DA79B7153")]
+	public class SepiaEffect : IGraphicsEffect, IGraphicsEffectSource, IGraphicsEffectD2D1Interop
+	{
+		private string _name = "SepiaEffect";
+		private Guid _id = new Guid("3A1AF410-5F1D-4DBE-84DF-915DA79B7153");
+
+		public string Name
+		{
+			get => _name;
+			set => _name = value;
+		}
+
+		public float Intensity { get; set; } = 0.5f;
+
+		public IGraphicsEffectSource Source { get; set; }
+
+		public Guid GetEffectId() => _id;
+
+		public void GetNamedPropertyMapping(string name, out uint index, out GraphicsEffectPropertyMapping mapping)
+		{
+			switch (name)
+			{
+				case "Intensity":
+					{
+						index = 0;
+						mapping = GraphicsEffectPropertyMapping.Direct;
+						break;
+					}
+				default:
+					{
+						index = 0xFF;
+						mapping = (GraphicsEffectPropertyMapping)0xFF;
+						break;
+					}
+			}
+		}
+
+		public object GetProperty(uint index)
+		{
+			switch (index)
+			{
+				case 0:
+					return Intensity;
+				default:
+					return null;
+			}
+		}
+
+		public uint GetPropertyCount() => 1;
+		public IGraphicsEffectSource GetSource(uint index) => Source;
+		public uint GetSourceCount() => 1;
+	}
+}
